Fix:noticeのスタイルを修正
