+++ src/components/organisms/Notification.tsx
@@ -1,7 +1,9 @@
 import React, { FC, Fragment } from "react";
 import { View, ScrollView, Text, ActivityIndicator } from "react-native";
 import { Image } from "react-native-elements";
-import { styles } from "../../styles/notification";
+import { StyleSheet } from "react-native";
+import { deviceWidth } from "../../utilities/dimensions";
+import { baseColor } from "../../styles/thema/colors";
 import NotificationText from "../atoms/NotificationText";
 
 type Props = {
@@ -11,8 +13,8 @@ type Props = {
 
 const Notification: FC<Props> = ({ notificationDataList }) => {
   return (
-    <ScrollView>
-      <View style={styles.container}>
+    <ScrollView style={styles.container}>
+      <View style={styles.box}>
         {notificationDataList.length !== 0 &&
         notificationDataList !== undefined ? (
           notificationDataList.map((item, index) => (
@@ -55,5 +57,48 @@ const Notification: FC<Props> = ({ notificationDataList }) => {
     </ScrollView>
   );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: baseColor.base,
+  },
+  box: {
+    alignItems: "center",
+  },
+  wrapper: {
+    flexDirection: "row",
+    padding: 15,
+  },
+  userImage: {
+    width: 52,
+    height: 52,
+    borderRadius: 180,
+  },
+  photoImage: {
+    width: 72,
+    height: 72,
+  },
+  border: {
+    width: deviceWidth,
+    borderBottomWidth: 0.5,
+    borderBottomColor: "#808080",
+  },
+  noNotification: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    width: deviceWidth / 1.3,
+    paddingBottom: 30,
+  },
+  noNotificationText: {
+    fontSize: 22,
+    color: "#fff",
+  },
+  supplement: {
+    marginTop: 30,
+    color: "#ddd",
+  },
+});
 
 export default Notification;
